Parse Twilio configuration from config file and command line.
Note: doesn't actually *do* anything with the Twilio configuration yet.

Authentication info goes in the config file (by default, ~/.wwdcc).
SMS from/to phone numbers are specified on the command line. The
authentication info cannot be specified on the command line as
it would be visible to anyone who runs 'ps'.