Fix a bug introduced by the global exception catching change.
Basically, we don't want to log and notify when catching ExitSuccess;
in fact, we'd rather not catch that exception at all, but I couldn't
figure out the pattern-matching mojo to pull that off, so I used the
"catches" function to hack something together. It works fine, it's just
a bit verbose.

See comments in Main.hs for details on the implementation.